Create a thumbnail window with a fitting colormap/visualid
This ensures that the window fits the OpenGL context. Taken from
> http://xcb.freedesktop.org/opengl/

The window is created with the size of the parent window, so that the
pixmap has already a certain size. Then, update() is called to render it
once. This speeds up showing of clients on the first key press a lot.