Improve call locations of update() and reset()
Not every call to update() requires a call to reset().
The call to update() in the c'tor is superfluous since a call to
update() is made on show().
Calls to update() and reset() for _NET_CLIENT_LIST_STACKING property
notify events are only necessary when the thumbnails are currently
shown.
All this eliminates the to check for the _active flag in update() and
reset().