Add default application icon as C source
Picture is taken from here:
> https://commons.wikimedia.org/wiki/File:Application-default-icon.svg

Idea from Openbox (obrender/theme.c) which (incorrectly) sets a default
window icon if neither NET_WM_ICON nor WM_ICON is present.

The SVG is converted with GIMP to a C source file. Since the blue and
red channels are swapped (ABGR) correction to ARBG is necessary. To
avoid doing this every time a default icon is necessary, we'll just
compute the picture in advance and store is as an array of bytes again,
which can be compiled in. Voilà, fast in-binary pictures.